filename of index of collection with only one document sectionsplit: https://github.com/metanorma/metanorma/issues/320
